+++ src/main/ipc/filesystem/filesystem-source-control-ai-targets.ts
@@ -175,11 +175,3 @@ export function getLocalTextGenerationTarget(
     ...(env ? { env } : {})
   }
 }
-
-export function validateFullGitObjectId(value: string, label: string): string {
-  const pattern = /^(?:[0-9a-fA-F]{40}|[0-9a-fA-F]{64})$/
-  if (!pattern.test(value)) {
-    throw new Error(`${label} must be a full git object id`)
-  }
-  return value
-}
