+++ src/renderer/src/runtime/sync-runtime-graph/mobile-session-editor-tabs.ts
@@ -9,7 +9,7 @@ import {
   isFileActiveEditorSurface,
   isMobileFileDiffSource,
   isUnifiedTabActiveInActiveGroup
-} from './mobile-session-tab-helpers'
+} from './mobile-session-surfaces'
 
 export function buildMobileMarkdownTab(
   inputs: MobileSessionWorktreeInputs,
