+++ src/renderer/src/store/slices/agent-status-launch-config.ts
@@ -7,9 +7,10 @@ import {
 } from '../../../../shared/agent-session-resume'
 import type {
   AgentLaunchConfigRegistryEntry,
-  AgentLaunchConfigRegistrationMetadata
+  AgentLaunchConfigRegistrationMetadata,
+  AgentLaunchConfigStatusMetadata
 } from './agent-status-contract'
-import { getLeafIdFromPaneKey, getTabIdFromPaneKey } from './agent-status-pane-helpers'
+import { getLeafIdFromPaneKey, getTabIdFromPaneKey } from './agent-status-pane-key-tab-binding'
 import { launchConfigsEqual } from './agent-status-recovery-equivalence'
 
 export function normalizeLaunchConfigRegistrationMetadata(
@@ -138,5 +139,25 @@ export function getLaunchConfigForEntry(
       entry.providerSession
     )
     ? sleepingRecord.launchConfig
+    : undefined
+}
+
+export function getLaunchConfigForStatusMetadata(
+  state: AppState,
+  metadata: AgentLaunchConfigStatusMetadata
+): SleepingAgentLaunchConfig | undefined {
+  const registryEntry = state.agentLaunchConfigByPaneKey[metadata.paneKey]
+  return registryEntryMatchesStatus({
+    entry: registryEntry,
+    paneKey: metadata.paneKey,
+    agentType: metadata.agentType,
+    tabId: metadata.tabId ?? getTabIdFromPaneKey(metadata.paneKey) ?? undefined,
+    terminalHandle: metadata.terminalHandle,
+    launchToken: metadata.launchToken,
+    providerSession: metadata.providerSession,
+    existingProviderSession: metadata.existingProviderSession,
+    providerSessionChanged: metadata.providerSessionChanged ?? false
+  })
+    ? registryEntry?.launchConfig
     : undefined
 }
