+++ src/renderer/src/store/slices/agent-status-live-actions.ts
@@ -16,7 +16,7 @@ import {
   agentStatusTabAlreadyHasProtectedOrGeneratedTitle,
   getTabIdFromPaneKey,
   isRecentlyClosedAgentStatusTab
-} from './agent-status-pane-helpers'
+} from './agent-status-pane-key-tab-binding'
 import {
   getAgentRowGeneratedTitleText,
   getOrcaDispatchTaskId,
