+++ src/renderer/src/store/slices/agent-status-orchestration-context.ts
@@ -1,0 +1,49 @@
+import type { AgentStatusOrchestrationContext } from '../../../../shared/agent-status-types'
+
+export function orchestrationContextsEqual(
+  a: AgentStatusOrchestrationContext,
+  b: AgentStatusOrchestrationContext
+): boolean {
+  return (
+    a.taskId === b.taskId &&
+    a.dispatchId === b.dispatchId &&
+    a.dispatchStatus === b.dispatchStatus &&
+    a.taskTitle === b.taskTitle &&
+    a.displayName === b.displayName &&
+    a.parentTerminalHandle === b.parentTerminalHandle &&
+    a.parentPaneKey === b.parentPaneKey &&
+    a.coordinatorHandle === b.coordinatorHandle &&
+    a.orchestrationRunId === b.orchestrationRunId
+  )
+}
+
+export function orchestrationMapsEqual(
+  a: Record<string, AgentStatusOrchestrationContext>,
+  b: Record<string, AgentStatusOrchestrationContext>
+): boolean {
+  const aKeys = Object.keys(a)
+  const bKeys = Object.keys(b)
+  if (aKeys.length !== bKeys.length) {
+    return false
+  }
+  return aKeys.every((key) => b[key] !== undefined && orchestrationContextsEqual(a[key]!, b[key]!))
+}
+
+export function mergeCurrentOrchestrationContext(
+  existing: AgentStatusOrchestrationContext | undefined,
+  current: AgentStatusOrchestrationContext
+): AgentStatusOrchestrationContext {
+  if (!existing) {
+    return current
+  }
+  const sameDispatch =
+    existing.taskId === current.taskId && existing.dispatchId === current.dispatchId
+  if (!sameDispatch) {
+    return current
+  }
+  if (current.dispatchStatus !== undefined) {
+    return orchestrationContextsEqual(existing, current) ? existing : current
+  }
+  const merged = { ...existing, ...current }
+  return orchestrationContextsEqual(existing, merged) ? existing : merged
+}
