+++ src/renderer/src/store/slices/agent-status-pane-keyed-records.ts
@@ -1,0 +1,76 @@
+export const RECENTLY_CLOSED_AGENT_STATUS_TAB_IDS_MAX = 1024
+export const RECENTLY_RETIRED_AGENT_STATUS_PANE_KEYS_MAX = 1024
+
+// delete-then-set for LRU recency, then evict oldest keys past the cap (Record iterates
+// insertion order); safe because a status for a tab closed >MAX tabs ago cannot still arrive.
+export function boundRecentlyClosedAgentStatusTabIds(
+  existing: Record<string, true>,
+  tabId: string
+): Record<string, true> {
+  const next: Record<string, true> = {}
+  for (const key of Object.keys(existing)) {
+    if (key !== tabId) {
+      next[key] = true
+    }
+  }
+  next[tabId] = true
+  const keys = Object.keys(next)
+  if (keys.length > RECENTLY_CLOSED_AGENT_STATUS_TAB_IDS_MAX) {
+    for (const stale of keys.slice(0, keys.length - RECENTLY_CLOSED_AGENT_STATUS_TAB_IDS_MAX)) {
+      delete next[stale]
+    }
+  }
+  return next
+}
+
+export function boundRecentlyRetiredAgentStatusPaneKeys(
+  existing: Record<string, true>,
+  paneKeys: readonly string[]
+): Record<string, true> {
+  const additions = new Set(paneKeys)
+  const next: Record<string, true> = {}
+  for (const key of Object.keys(existing)) {
+    if (!additions.has(key)) {
+      next[key] = true
+    }
+  }
+  for (const paneKey of additions) {
+    next[paneKey] = true
+  }
+  const keys = Object.keys(next)
+  for (const stale of keys.slice(0, -RECENTLY_RETIRED_AGENT_STATUS_PANE_KEYS_MAX)) {
+    delete next[stale]
+  }
+  return next
+}
+
+export function movePaneKeyedRecord<T>(
+  record: Record<string, T>,
+  fromPaneKey: string,
+  toPaneKey: string,
+  transform: (value: T) => T = (value) => value
+): Record<string, T> {
+  const value = record[fromPaneKey]
+  if (value === undefined || fromPaneKey === toPaneKey) {
+    return record
+  }
+  const next = { ...record }
+  delete next[fromPaneKey]
+  next[toPaneKey] = transform(value)
+  return next
+}
+
+export function removePaneKeys<T>(
+  record: Record<string, T>,
+  paneKeys: ReadonlySet<string>
+): Record<string, T> {
+  const matchingKeys = Object.keys(record).filter((key) => paneKeys.has(key))
+  if (matchingKeys.length === 0) {
+    return record
+  }
+  const next = { ...record }
+  for (const key of matchingKeys) {
+    delete next[key]
+  }
+  return next
+}
