+++ src/renderer/src/store/slices/agent-status-worktree-drop-actions.ts
@@ -1,13 +1,13 @@
 import type { DropAgentStatusByWorktreeOptions, RetainedAgentEntry } from './agent-status-contract'
 import type { AgentStatusSlice } from './agent-status-slice-contract'
 import type { AgentStatusRuntime } from './agent-status-runtime'
-import { pruneMigrationUnsupportedEntries } from './agent-status-map-helpers'
+import { pruneMigrationUnsupportedEntries } from './agent-status-migration-unsupported-entries'
 import {
   normalizePaneKeySet,
   paneKeyMatchesAnyTabPrefix,
   retainedAgentEntryFromLive,
   shouldReplaceRetainedWithLive
-} from './agent-status-pane-helpers'
+} from './agent-status-pane-key-tab-binding'
 
 export function createAgentStatusWorktreeDropActions(
   runtime: AgentStatusRuntime
