+++ src/renderer/src/store/slices/ui/ui-slice-agent-notification-acknowledgement.ts
@@ -53,10 +53,3 @@ export function latestAgentTurnTimestamp(entry: {
   }
   return latest
 }
-
-let agentSendTargetModeInstanceCounter = 0
-
-export function createAgentSendTargetModeInstanceId(): string {
-  agentSendTargetModeInstanceCounter += 1
-  return `${Date.now()}:${agentSendTargetModeInstanceCounter}`
-}
