+++ src/renderer/src/store/slices/ui/ui-slice-contextual-tour-progression.ts
